requirements.txt
schedule==1.1.0
pystray==0.17.3
selenium==3.141.0
schedule==1.1.0
Pillow==9.1.0
